+++ pang/src/symbol/terminals/bits.rs
@@ -4,7 +4,7 @@ use std::{any::Any, hash::Hasher, sync::Arc};
 
 use serde::{Deserialize, Serialize};
 
-use crate::symbol::{DecodeResult, SharedState, Symbol, terminals::TerminalKind};
+use crate::symbol::{Symbol, terminals::TerminalKind};
 
 /// Bits terminal, not fully implemented yet.
 #[derive(Clone, PartialEq, Debug, Eq, Hash, Serialize, Deserialize)]
@@ -43,14 +43,6 @@ impl TerminalKind for BitsTerminal {
         } else {
             Ok(self.value[..self.size].to_vec())
         }
-    }
-
-    fn parse<'a>(
-        &self,
-        _input: &'a [u8],
-        _state: &SharedState,
-    ) -> DecodeResult<'a, Arc<dyn TerminalKind>> {
-        todo!("Implement it later.")
     }
 
     fn as_any(&self) -> &dyn Any {